+++ python/lsst/ts/criopy/m1m3/offsets_widget.py
@@ -235,7 +235,7 @@ class OffsetsWidget(QWidget):
         row += 1
         self.dir_pad = DirectionPadWidget()
         self.dir_pad.setEnabled(False)
-        self.dir_pad.positionChanged.connect(self._positionChanged)
+        self.dir_pad.positionChanged.connect(self._position_changed)
         data_layout.addWidget(self.dir_pad, row, 1, 3, 6)
 
         self.preclipped = createForces()
@@ -243,11 +243,12 @@ class OffsetsWidget(QWidget):
         self.measured = createForces()
 
         row += 3
+        data_layout.addWidget(QLabel())
         for i in range(6):
             text = (
-                f"Force {chr(ord('W') + i)}"
+                f"Force {chr(ord('W') + i + 1)}"
                 if i < 3
-                else f"Moment {chr(ord('W') + i - 3)}"
+                else f"Moment {chr(ord('W') + i - 2)}"
             )
             data_layout.addWidget(QLabel(f"<b>{text}</b>"))
 
@@ -278,24 +279,24 @@ class OffsetsWidget(QWidget):
             col += 1
 
         row += 1
-        self.offsetForces = QPushButton("Apply offset forces")
-        self.offsetForces.setEnabled(False)
-        self.offsetForces.clicked.connect(self._applyOffsetForces)
-        data_layout.addWidget(self.offsetForces, row, 1, 1, 3)
+        self.offset_forces = QPushButton("Apply offset forces")
+        self.offset_forces.setEnabled(False)
+        self.offset_forces.clicked.connect(self._apply_offset_forces)
+        data_layout.addWidget(self.offset_forces, row, 1, 1, 3)
 
-        self.clearOffsetForcesButton = QPushButton("Reset forces")
-        self.clearOffsetForcesButton.setEnabled(False)
-        self.clearOffsetForcesButton.clicked.connect(self._clearOffsetForces)
-        data_layout.addWidget(self.clearOffsetForcesButton, row, 4, 1, 3)
+        self.clear_offset_forces_button = QPushButton("Reset forces")
+        self.clear_offset_forces_button.setEnabled(False)
+        self.clear_offset_forces_button.clicked.connect(self._clear_offset_forces)
+        data_layout.addWidget(self.clear_offset_forces_button, row, 4, 1, 3)
 
         layout.addStretch()
 
-        self.m1m3.hardpointActuatorData.connect(self._hardpointActuatorDataCallback)
-        self.m1m3.imsData.connect(self._imsDataCallback)
-        self.m1m3.preclippedOffsetForces.connect(self._preclippedOffsetForces)
-        self.m1m3.appliedOffsetForces.connect(self._appliedOffsetForces)
-        self.m1m3.forceActuatorData.connect(self._forceActuatorCallback)
-        self.m1m3.detailedState.connect(self._detailedStateCallback)
+        self.m1m3.hardpointActuatorData.connect(self._hardpoint_actuator_data_callback)
+        self.m1m3.imsData.connect(self._ims_data_callback)
+        self.m1m3.preclippedOffsetForces.connect(self._preclipped_offset_forces)
+        self.m1m3.appliedOffsetForces.connect(self._applied_offset_forces)
+        self.m1m3.forceActuatorData.connect(self._force_actuator_callback)
+        self.m1m3.detailedState.connect(self._detailed_state_callback)
 
     async def moveMirror(self, **kwargs: typing.Any) -> None:
         """Move mirror. Calls positionM1M3 command.
@@ -336,7 +337,7 @@ class OffsetsWidget(QWidget):
         for k, v in targets.items():
             getattr(self, "target_" + k).setValue(v / self._getScale(k))
 
-    def getForceOffsets(self) -> dict[str, float]:
+    def get_force_offsets(self) -> dict[str, float]:
         """Return current offset forces (from forceOffsets_ box).
 
         Returns
@@ -372,21 +373,21 @@ class OffsetsWidget(QWidget):
         self.dir_pad.set_position(args[p] for p in self.POSITIONS)
 
     @asyncSlot()
-    async def _applyOffsetForces(self) -> None:
+    async def _apply_offset_forces(self) -> None:
         await command(
             self,
             self.m1m3.remote.cmd_applyOffsetForcesByMirrorForce,
-            **self.getForceOffsets(),
+            **self.get_force_offsets(),
         )
 
     @asyncSlot()
-    async def _clearOffsetForces(self) -> None:
-        await command(self, self.m1m3.remote.cmd_clearOffsetForces)
+    async def _clear_offset_forces(self) -> None:
+        await command(self, self.m1m3.remote.cmd_clear_offset_forces)
         for f in self.FORCES:
             getattr(self, "forceOffsets_" + f).setValue(0)
 
     @asyncSlot()
-    async def _positionChanged(self, offsets: list[float]) -> None:
+    async def _position_changed(self, offsets: list[float]) -> None:
         args = {}
         for i in range(6):
             args[self.POSITIONS[i]] = offsets[i]
@@ -409,7 +410,7 @@ class OffsetsWidget(QWidget):
             v.setValue(getattr(self.__hp_data, k) - getattr(self.__ims_data, k))
 
     @Slot()
-    def _hardpointActuatorDataCallback(self, data: BaseMsgType) -> None:
+    def _hardpoint_actuator_data_callback(self, data: BaseMsgType) -> None:
         self.__fill_row(self.hp_position, data)
         self.__fill_row(self.hp_forces, data)
         self.__fill_array(self.hp_measured_forces, data)
@@ -420,29 +421,29 @@ class OffsetsWidget(QWidget):
         self.__update_diffs()
 
     @Slot()
-    def _imsDataCallback(self, data: BaseMsgType) -> None:
+    def _ims_data_callback(self, data: BaseMsgType) -> None:
         self.__fill_row(self.ims_position, data)
         self.__ims_data = data
         self.__update_diffs()
 
     @Slot()
-    def _preclippedOffsetForces(self, data: typing.Any) -> None:
+    def _preclipped_offset_forces(self, data: typing.Any) -> None:
         self.__fill_row(self.preclipped, data)
 
     @Slot()
-    def _appliedOffsetForces(self, data: typing.Any) -> None:
+    def _applied_offset_forces(self, data: typing.Any) -> None:
         self.__fill_row(self.applied, data)
 
     @Slot()
-    def _forceActuatorCallback(self, data: typing.Any) -> None:
+    def _force_actuator_callback(self, data: typing.Any) -> None:
         self.__fill_row(self.measured, data)
 
     @Slot()
-    def _detailedStateCallback(self, data: typing.Any) -> None:
+    def _detailed_state_callback(self, data: typing.Any) -> None:
         enabled = data.detailedState == DetailedStates.ACTIVEENGINEERING
 
         self.move_mirror_button.setEnabled(enabled)
         self.stop_mirror_button.setEnabled(enabled)
         self.dir_pad.setEnabled(enabled)
-        self.offsetForces.setEnabled(enabled)
-        self.clearOffsetForcesButton.setEnabled(enabled)
+        self.offset_forces.setEnabled(enabled)
+        self.clear_offset_forces_button.setEnabled(enabled)